Enable pane titles in one tmux window
This shows the pane title for the docker window. Previously, it was not
clear which environment was displayed in the panes.

Using the `pre` and `pre_window` hooks somehow didn't work as expected.
But tmux accepts targets via the -t parameter, so using the docker window
as target works fine.